refactor: motion and swisseph_setup become packages
motion.py -> motion/state.py, swisseph_setup.py -> swisseph_setup/download.py.

swisseph_setup keeps its package name because `python -m
kerykeion.swisseph_setup` is documented in the README, the release notes
and tests/conftest.py; the CLI now reaches it through __main__.py.

argparse gets an explicit prog=: under `python -m <package>` it would
otherwise derive the program name from __main__.py, silently changing
what --help prints.
