+++ kerykeion/motion/state.py
