+++ kerykeion/swisseph_setup/__main__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""Entry point for ``python -m kerykeion.swisseph_setup``."""
+
+from .download import main
+
+if __name__ == "__main__":
+    main()
